requirements.txt
torch>=1.10
transformers>=4.26
pandas
numpy
pyyaml
scikit-learn
num2words
matplotlib
tqdm
# Untuk notebook dan EDA (opsional)
notebook
ipykernel
# Untuk evaluasi/serving ONNX (opsional)
onnx
onnxruntime
